requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.15.0
gspread>=5.10.0
google-auth>=2.17.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
